+++ packages/stream_video/lib/src/coordinator/open_api/coordinator_client_open_api.dart
@@ -124,6 +124,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
                 connectionId: event.connectionId,
                 closeCode: event.closeCode,
                 closeReason: event.closeReason,
+                apiError: event.apiError,
               );
             }
             _events.emit(event);